% Auto-generated: do not edit by hand
\name{dashEditor}

\alias{dashEditor}

\title{DashEditor component}

\description{
EditorComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashEditor(id=NULL, dataSources=NULL, fig=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{dataSources}{Named list. The layout of the  components displayed inside the grid.}

\item{fig}{Named list. The layout of the  components displayed inside the grid.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

